+++ packages/selector-specificity/test/test.mjs
@@ -106,6 +106,7 @@ assert.deepEqual(calculate(':is(#a + #a + #a, #b + #b + #b.b)'), { a: 3, b: 1, c
 
 assert.deepEqual(calculate('::view-transition'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-group(foo)'), { a: 0, b: 0, c: 1 });
+assert.deepEqual(calculate('::view-transition-group()'), { a: 0, b: 0, c: 0 });
 assert.deepEqual(calculate('::view-transition-group(*)'), { a: 0, b: 0, c: 0 });
 assert.deepEqual(calculate('::view-transition-group(*.foo)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-group(foo.foo)'), { a: 0, b: 0, c: 1 });
@@ -120,6 +121,7 @@ assert.deepEqual(calculate('::view-transition-image-pair(.foo.fooz)'), { a: 0, b
 assert.deepEqual(calculate('::view-transition-image-pair(foo.foo.fooz)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-image-pair(*.foo.fooz)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-old(foo)'), { a: 0, b: 0, c: 1 });
+assert.deepEqual(calculate('::view-transition-old()'), { a: 0, b: 0, c: 0 });
 assert.deepEqual(calculate('::view-transition-old(*)'), { a: 0, b: 0, c: 0 });
 assert.deepEqual(calculate('::view-transition-old(*.foo)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-old(foo.foo)'), { a: 0, b: 0, c: 1 });
@@ -127,6 +129,7 @@ assert.deepEqual(calculate('::view-transition-old(.foo.fooz)'), { a: 0, b: 0, c:
 assert.deepEqual(calculate('::view-transition-old(foo.foo.fooz)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-old(*.foo.fooz)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-new(foo)'), { a: 0, b: 0, c: 1 });
+assert.deepEqual(calculate('::view-transition-new()'), { a: 0, b: 0, c: 0 });
 assert.deepEqual(calculate('::view-transition-new(*)'), { a: 0, b: 0, c: 0 });
 assert.deepEqual(calculate('::view-transition-new(*.foo)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-new(foo.foo)'), { a: 0, b: 0, c: 1 });
@@ -134,14 +137,18 @@ assert.deepEqual(calculate('::view-transition-new(.foo.fooz)'), { a: 0, b: 0, c:
 assert.deepEqual(calculate('::view-transition-new(foo.foo.fooz)'), { a: 0, b: 0, c: 1 });
 assert.deepEqual(calculate('::view-transition-new(*.foo.fooz)'), { a: 0, b: 0, c: 1 });
 
-assert.deepEqual(calculate(':active-view-transition(*)'), { a: 0, b: 1, c: 0 });
-assert.deepEqual(calculate(':active-view-transition(foo)'), { a: 0, b: 2, c: 0 });
-assert.deepEqual(calculate(':active-view-transition(.foo)'), { a: 0, b: 2, c: 0 });
-assert.deepEqual(calculate(':active-view-transition(.foo.bar)'), { a: 0, b: 2, c: 0 });
-assert.deepEqual(calculate(':active-view-transition(.foo .bar)'), { a: 0, b: 2, c: 0 });
-assert.deepEqual(calculate(':active-view-transition(foo, bar)'), { a: 0, b: 2, c: 0 });
+assert.deepEqual(calculate(':active-view-transition'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(foo)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(.foo)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(.foo.bar)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(.foo .bar)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(foo, bar)'), { a: 0, b: 1, c: 0 });
 
-// Invalid CSS, must be either `*` or a list of one or more custom idents
+// Invalid CSS.
 // We should still calculate some specificity.
-assert.deepEqual(calculate(':active-view-transition(*, bar)'), { a: 0, b: 2, c: 0 });
-assert.deepEqual(calculate(':active-view-transition(*, *)'), { a: 0, b: 2, c: 0 });
+assert.deepEqual(calculate(':active-view-transition(*)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition(*, bar)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition(*, *)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(*)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(*, bar)'), { a: 0, b: 1, c: 0 });
+assert.deepEqual(calculate(':active-view-transition-type(*, *)'), { a: 0, b: 1, c: 0 });